requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
jupyter
notebook
torch
sentence-transformers
faiss-cpu
transformers
streamlit
pdfplumber
pytest
